autoclipper-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
requests==2.32.3
feedparser==6.0.11
redis==5.2.1
rq==1.16.2
apscheduler==3.10.4
yt-dlp
openai-whisper
ffmpeg-python
google-generativeai
groq
